requirements.txt
Flask==2.3.3
selenium==4.15.2
pandas==2.3.1
numpy==2.3.1
xlsxwriter==3.1.9
requests==2.31.0
python-dotenv==1.0.0
aiohttp